requirements.txt
Flask==2.3.6
gunicorn==22.1.0
openai==1.32.0
gspread==5.9.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
